// examples/vector-search/bunvex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `bunvex dev`.
 * @module
 */

import type * as embed from "../embed.js";
import type * as foods from "../foods.js";
import type * as movies from "../movies.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "bunvex/server";

declare const fullApi: ApiFromModules<{
  embed: typeof embed;
  foods: typeof foods;
  movies: typeof movies;
}>;

/**
 * A utility for referencing bunvex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing bunvex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};